Moving away from legacy code now using core profile internally
 * minimum GFX card requirement to run Fragmentarium: OpenGL 3.3 core
   (precompiled release versions use 4.1 compatibility for Windows,
    4.1 core for Mac and 4.5 core for Linux)
 * better OS vs GL profile handling
 * spline shader works for all versions and profiles available for testing
 * discover/estimate the GLSL capabilities and display in log window
 * fixed texture fail on windows version caused by latin1 vs 8bit char for file name

[quote]
GLSLangSpec.4.50
Version 110 of the language does not require shaders to include the #version directive,
and shaders that do not include a #version directive will be treated as targeting version 110

Shaders that specify #version 100 will be treated as targeting version 1.00 of the OpenGL ES Shading Language.
Shaders that specify #version 300 will be treated as targeting version 3.00 of the OpenGL ES Shading Language.
Shaders that specify #version 310 will be treated as targeting version 3.10 of the OpenGL ES Shading Language.
[/quote]